Token Factory DAOs (#731)

+++ justfile
@@ -19,10 +19,16 @@ gen-schema:
 integration-test: deploy-local workspace-optimize
 	RUST_LOG=info CONFIG={{orc_config}} cargo integration-test
 
+test-tube:
+    cargo test --features "test-tube"
+
+test-tube-dev: workspace-optimize
+    cargo test --features "test-tube"
+
 integration-test-dev test_name="":
 	SKIP_CONTRACT_STORE=true RUST_LOG=info CONFIG='{{`pwd`}}/ci/configs/cosm-orc/local.yaml' cargo integration-test {{test_name}}
 
-bootstrap-dev: deploy-local workspace-optimize-arm
+bootstrap-dev: deploy-local workspace-optimize
 	RUST_LOG=info CONFIG={{orc_config}} cargo run bootstrap-env
 
 deploy-local: download-deps
@@ -39,7 +45,7 @@ deploy-local: download-deps
 		-p 26657:26657 \
 		-p 9090:9090 \
 		--mount type=volume,source=junod_data,target=/root \
-		ghcr.io/cosmoscontracts/juno:v11.0.0 /opt/setup_and_run.sh {{test_addrs}}
+		ghcr.io/cosmoscontracts/juno:v15.0.0 /opt/setup_and_run.sh {{test_addrs}}
 
 download-deps:
 	mkdir -p artifacts target
@@ -47,20 +53,20 @@ download-deps:
 	wget https://github.com/CosmWasm/cw-plus/releases/latest/download/cw4_group.wasm -O artifacts/cw4_group.wasm
 	wget https://github.com/CosmWasm/cw-nfts/releases/latest/download/cw721_base.wasm -O artifacts/cw721_base.wasm
 
-optimize:
-	cargo install cw-optimizoor || true
-	cargo cw-optimizoor Cargo.toml
-
 workspace-optimize:
-	docker run --rm -v "$(pwd)":/code \
-		--mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
-		--mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-		--platform linux/amd64 \
-		cosmwasm/workspace-optimizer:0.12.13
-
-workspace-optimize-arm:
-    docker run --rm -v "$(pwd)":/code \
-        --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
-        --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-        --platform linux/arm64 \
-        cosmwasm/workspace-optimizer-arm64:0.12.13
+    #!/bin/bash
+    if [[ $(uname -m) == 'arm64' ]]; then docker run --rm -v "$(pwd)":/code \
+            --mount type=volume,source="$(basename "$(pwd)")_cache",target=/target \
+            --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+            --platform linux/arm64 \
+            cosmwasm/workspace-optimizer-arm64:0.13.0; \
+    elif [[ $(uname -m) == 'aarch64' ]]; then docker run --rm -v "$(pwd)":/code \
+            --mount type=volume,source="$(basename "$(pwd)")_cache",target=/target \
+            --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+            --platform linux/arm64 \
+            cosmwasm/workspace-optimizer-arm64:0.13.0; \
+    elif [[ $(uname -m) == 'x86_64' ]]; then docker run --rm -v "$(pwd)":/code \
+            --mount type=volume,source="$(basename "$(pwd)")_cache",target=/target \
+            --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
+            --platform linux/amd64 \
+            cosmwasm/workspace-optimizer:0.13.0; fi
